+++ src/client/components/ac_signin/ac_signin.js
@@ -50,7 +50,7 @@ Template.ac_signin.helpers({
     onlyEmailAddress(){
         const acName = this.AC.options.acName();
         const acInstance = AccountsCore.getInstance( acName );
-        assert( acInstance && acInstance instanceof AccountsCore.acAccount, 'expects an instance of AccountsCore.acAccount, got '+acInstance );
+        assert( acInstance && acInstance instanceof AccountsCore.Account, 'expects an instance of AccountsCore.Account, got '+acInstance );
         return acInstance.opts().haveEmailAddress() !== AccountsCore.C.Identifier.NONE && acInstance.opts().haveUsername() === AccountsCore.C.Identifier.NONE;
     },
 
@@ -58,7 +58,7 @@ Template.ac_signin.helpers({
     onlyEmailAddress(){
         const acName = this.AC.options.acName();
         const acInstance = AccountsCore.getInstance( acName );
-        assert( acInstance && acInstance instanceof AccountsCore.acAccount, 'expects an instance of AccountsCore.acAccount, got '+acInstance );
+        assert( acInstance && acInstance instanceof AccountsCore.Account, 'expects an instance of AccountsCore.Account, got '+acInstance );
         return acInstance.opts().haveEmailAddress() === AccountsCore.C.Identifier.NONE && acInstance.opts().haveUsername() !== AccountsCore.C.Identifier.NONE;
     },